Fixed finance DB tests.

+++ src/finance_db/table_definitions.cc
@@ -9,7 +9,7 @@ auto Asset() -> const sqldb::TableDefinition& {
   static const auto table_definition =
       sqldb::TableDefinition{.table = "Asset",
                              .columns = {{.column = "id",
-                                          .data_type = sqldb::DataType::kInt,
+                                          .data_type = sqldb::DataType::kInt64,
                                           .primary_key = true,
                                           .auto_increment = true,
                                           .unique = true},
@@ -24,7 +24,7 @@ auto Symbol() -> const sqldb::TableDefinition& {
       .table = "Symbol",
       .columns = {
           {.column = "id",
-           .data_type = sqldb::DataType::kInt,
+           .data_type = sqldb::DataType::kInt64,
            .primary_key = true,
            .auto_increment = true,
            .unique = true},
@@ -32,10 +32,10 @@ auto Symbol() -> const sqldb::TableDefinition& {
            .data_type = sqldb::DataType::kString,
            .unique = true},
           {.column = "base_asset_id",
-           .data_type = sqldb::DataType::kInt,
+           .data_type = sqldb::DataType::kInt64,
            .foreign_key = {{.table = "Asset", .column = "id"}}},
           {.column = "quote_asset_id",
-           .data_type = sqldb::DataType::kInt,
+           .data_type = sqldb::DataType::kInt64,
            .foreign_key = {{.table = "Asset", .column = "id"}}},
           {.column = "min_base_amount", .data_type = sqldb::DataType::kDouble},
           {.column = "min_quote_amount", .data_type = sqldb::DataType::kDouble},
@@ -49,9 +49,9 @@ auto SymbolPriceTick() -> const sqldb::TableDefinition& {
       .table = "SymbolPriceTick",
       .columns = {
           {.column = "symbol_id",
-           .data_type = sqldb::DataType::kInt,
+           .data_type = sqldb::DataType::kInt64,
            .foreign_key = {{.table = "Symbol", .column = "id"}}},
-          {.column = "time", .data_type = sqldb::DataType::kInt},
+          {.column = "time", .data_type = sqldb::DataType::kInt64},
           {.column = "buy_price", .data_type = sqldb::DataType::kDouble},
           {.column = "sell_price", .data_type = sqldb::DataType::kDouble}}};
   return table_definition;
